Start moving get_concatenated_data to Concatenator
